Added blogs delete method, refactoring

+++ src/blogs/blogs.service.ts
@@ -4,6 +4,7 @@ import { Blog, BlogModelType } from './schemas/blog.entity';
 import { BlogCreateDto } from './dto/blog.create.dto';
 import { BlogsRepository } from './blogs.repository';
 import { BlogUpdateDto } from './dto/blog.update.dto';
+import { BlogViewModel } from './schemas/blog.view';
 
 @Injectable()
 export class BlogsService {
@@ -13,9 +14,9 @@ export class BlogsService {
     private readonly blogsRepository: BlogsRepository,
   ) {}
 
-  async createBlog(createBlogDto: BlogCreateDto): Promise<Blog> {
+  async createBlog(createBlogDto: BlogCreateDto): Promise<BlogViewModel> {
     const blog = this.BlogModel.createBlog(createBlogDto, this.BlogModel);
-    return this.blogsRepository.save(blog);
+    return this.blogsRepository.createBlog(blog);
   }
 
   async updateBlog(id: string, updateBlogDto: BlogUpdateDto): Promise<Blog> {
@@ -31,7 +32,15 @@ export class BlogsService {
     return this.blogsRepository.save(blog);
   }
 
-  /*async deleteBlog(id: string): Promise<boolean> {
+  async deleteBlog(id: string): Promise<boolean> {
+    const blog = await this.blogsRepository.findBlog(id);
+
+    if (!blog) {
+      throw new InternalServerErrorException(
+        `Something went wrong during blog find operation`,
+      );
+    }
+
     return this.blogsRepository.deleteBlog(id);
-  }*/
+  }
 }
